Address more code review comments.
Also a minor bugfix.

(sapling split of 263d945d065bb1f788fd10ab3a8b5e7467bc46bd)

+++ src/python/twitter/pants/commands/goal.py
@@ -479,9 +479,10 @@ goal(
 
 def async_safe_rmtree(root):
   new_path = root + '.deletable.%f' % time.time()
-  os.rename(root, new_path)
-  with daemon.DaemonContext():
-    safe_rmtree(new_path)
+  if os.path.exists(root):
+    os.rename(root, new_path)
+    with daemon.DaemonContext():
+      safe_rmtree(new_path)
 
 goal(
   name='clean-all-async',
